update config location, get mouse events hooked up again
