+++ source/CoordinateConversion/CoordinateConversionLibrary/ViewModels/OutputCoordinateViewModel.cs
@@ -20,12 +20,9 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Xml.Serialization;
-using System.Xml;
-using System.IO;
 using CoordinateConversionLibrary.Models;
 using CoordinateConversionLibrary.Helpers;
 using CoordinateConversionLibrary.Views;
-using System.Windows;
 
 namespace CoordinateConversionLibrary.ViewModels
 {
@@ -38,8 +35,13 @@ namespace CoordinateConversionLibrary.ViewModels
             DeleteCommand = new RelayCommand(OnDeleteCommand);
             CopyCommand = new RelayCommand(OnCopyCommand);
 
+            // set default CoordinateGetter
+            coordinateGetter = new CoordinateGetBase();
+
             Mediator.Register(CoordinateConversionLibrary.Constants.AddNewOutputCoordinate, OnAddNewOutputCoordinate);
             Mediator.Register(CoordinateConversionLibrary.Constants.CopyAllCoordinateOutputs, OnCopyAllCoordinateOutputs);
+            Mediator.Register(CoordinateConversionLibrary.Constants.SetCoordinateGetter, OnSetCoordinateGetter);
+            Mediator.Register(CoordinateConversionLibrary.Constants.RequestOutputUpdate, OnOutputUpdate);
 
             //for testing without a config file, init a few sample items
             //OutputCoordinateList = new ObservableCollection<OutputCoordinateModel>();
@@ -57,11 +59,13 @@ namespace CoordinateConversionLibrary.ViewModels
 
             //DefaultFormatList.Add(new DefaultFormatModel { CType = CoordinateType.DD, DefaultNameFormatDictionary = new SerializableDictionary<string, string> { { "70.49N 40.32W", "Y0.0#N X0.0#E" }, { "70.49N,40.32W", "Y0.0#N,X0.0#E" } } });
 
-            configObserver = new PropertyObserver<CoordinateConversionLibraryConfig>(CoordinateConversionViewModel.AddInConfig)
+            configObserver = new PropertyObserver<CoordinateConversionLibraryConfig>(CoordinateConversionLibraryConfig.AddInConfig)
             .RegisterHandler(n => n.OutputCoordinateList, n=> RaisePropertyChanged(()=> OutputCoordinateList));
         }
 
         PropertyObserver<CoordinateConversionLibraryConfig> configObserver;
+
+        private CoordinateGetBase coordinateGetter;
 
         private void OnCopyAllCoordinateOutputs(object obj)
         {
@@ -72,7 +76,7 @@ namespace CoordinateConversionLibrary.ViewModels
             if (!string.IsNullOrWhiteSpace(inputCoordinate))
                 sb.AppendLine(inputCoordinate);
 
-            foreach (var output in CoordinateConversionViewModel.AddInConfig.OutputCoordinateList)
+            foreach (var output in CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList)
             {
                 sb.AppendLine(output.OutputCoordinate);
             }
@@ -91,7 +95,7 @@ namespace CoordinateConversionLibrary.ViewModels
             if (outputCoordItem == null)
                 return;
 
-            var dlg = new EditOutputCoordinateView(CoordinateConversionViewModel.AddInConfig.DefaultFormatList, GetInUseNames(), new OutputCoordinateModel() { CType = outputCoordItem.CType, Format = outputCoordItem.Format, Name = outputCoordItem.Name, SRName = outputCoordItem.SRName, SRFactoryCode = outputCoordItem.SRFactoryCode });
+            var dlg = new EditOutputCoordinateView(CoordinateConversionLibraryConfig.AddInConfig.DefaultFormatList, GetInUseNames(), new OutputCoordinateModel() { CType = outputCoordItem.CType, Format = outputCoordItem.Format, Name = outputCoordItem.Name, SRName = outputCoordItem.SRName, SRFactoryCode = outputCoordItem.SRFactoryCode });
 
             var vm = dlg.DataContext as EditOutputCoordinateViewModel;
             vm.WindowTitle = Properties.Resources.TitleAddNewOutputCoordinate;
@@ -110,15 +114,25 @@ namespace CoordinateConversionLibrary.ViewModels
                 outputCoordItem.SRFactoryCode = vm.OutputCoordItem.SRFactoryCode;
                 outputCoordItem.SRName = vm.OutputCoordItem.SRName;
 
-                CoordinateConversionViewModel.AddInConfig.OutputCoordinateList.Add(outputCoordItem);
-                Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.RequestOutputUpdate, null);
-                CoordinateConversionViewModel.AddInConfig.SaveConfiguration();
+                CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList.Add(outputCoordItem);
+                UpdateOutputs();
+                CoordinateConversionLibraryConfig.AddInConfig.SaveConfiguration();
             }
+        }
+
+        private void OnSetCoordinateGetter(object obj)
+        {
+            coordinateGetter = obj as CoordinateGetBase;
+        }
+
+        private void OnOutputUpdate(object obj)
+        {
+            UpdateOutputs();
         }
 
         private List<string> GetInUseNames()
         {
-            return CoordinateConversionViewModel.AddInConfig.OutputCoordinateList.Select(oc => oc.Name).ToList();
+            return CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList.Select(oc => oc.Name).ToList();
         }
 
         /// <summary>
@@ -126,7 +140,7 @@ namespace CoordinateConversionLibrary.ViewModels
         /// </summary>
         public ObservableCollection<OutputCoordinateModel> OutputCoordinateList
         {
-            get { return CoordinateConversionViewModel.AddInConfig.OutputCoordinateList; }
+            get { return CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList; }
         }
 
         #region relay commands
@@ -161,12 +175,12 @@ namespace CoordinateConversionLibrary.ViewModels
                 if (System.Windows.MessageBoxResult.Yes != System.Windows.MessageBox.Show(string.Format(Properties.Resources.FormattedRemove, name), Properties.Resources.LabelConfirmRemoval, System.Windows.MessageBoxButton.YesNo))
                     return;
 
-                foreach (var item in CoordinateConversionViewModel.AddInConfig.OutputCoordinateList)
+                foreach (var item in CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList)
                 {
                     if (item.Name == name)
                     {
-                        CoordinateConversionViewModel.AddInConfig.OutputCoordinateList.Remove(item);
-                        CoordinateConversionViewModel.AddInConfig.SaveConfiguration();
+                        CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList.Remove(item);
+                        CoordinateConversionLibraryConfig.AddInConfig.SaveConfiguration();
                         return;
                     }
                 }
@@ -179,7 +193,7 @@ namespace CoordinateConversionLibrary.ViewModels
 
             if(!string.IsNullOrWhiteSpace(name))
             {
-                foreach (var item in CoordinateConversionViewModel.AddInConfig.OutputCoordinateList)
+                foreach (var item in CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList)
                 {
                     if(item.Name == name)
                     {
@@ -198,7 +212,7 @@ namespace CoordinateConversionLibrary.ViewModels
             var outputCoordItem = GetOCMByName(obj as string);
             var InUseNames = GetInUseNames();
             InUseNames.Remove(outputCoordItem.Name);
-            var dlg = new EditOutputCoordinateView(CoordinateConversionViewModel.AddInConfig.DefaultFormatList, InUseNames, 
+            var dlg = new EditOutputCoordinateView(CoordinateConversionLibraryConfig.AddInConfig.DefaultFormatList, InUseNames, 
                 new OutputCoordinateModel() { CType = outputCoordItem.CType, 
                     Format = outputCoordItem.Format, 
                     Name = outputCoordItem.Name,
@@ -221,10 +235,10 @@ namespace CoordinateConversionLibrary.ViewModels
                     outputCoordItem.CType = type;
                 }
 
-                Mediator.NotifyColleagues(CoordinateConversionLibrary.Constants.RequestOutputUpdate, null);
+                UpdateOutputs();
             }
 
-            CoordinateConversionViewModel.AddInConfig.SaveConfiguration();
+            CoordinateConversionLibraryConfig.AddInConfig.SaveConfiguration();
         }
 
         #endregion
@@ -233,7 +247,7 @@ namespace CoordinateConversionLibrary.ViewModels
         {
             if (!string.IsNullOrEmpty(name))
             {
-                foreach (var item in CoordinateConversionViewModel.AddInConfig.OutputCoordinateList)
+                foreach (var item in CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList)
                 {
                     if (item.Name == name)
                     {
@@ -244,5 +258,159 @@ namespace CoordinateConversionLibrary.ViewModels
 
             return null;
         }
+
+        private void UpdateOutputs()
+        {
+            foreach (var output in CoordinateConversionLibraryConfig.AddInConfig.OutputCoordinateList)
+            {
+                var props = new Dictionary<string, string>();
+                string coord = string.Empty;
+
+                switch (output.CType)
+                {
+                    case CoordinateType.DD:
+                        CoordinateDD cdd;
+                        if (coordinateGetter.CanGetDD(output.SRFactoryCode, out coord) &&
+                            CoordinateDD.TryParse(coord, out cdd))
+                        {
+                            output.OutputCoordinate = cdd.ToString(output.Format, new CoordinateDDFormatter());
+                            var splits = output.Format.Split(new char[] { 'X' }, StringSplitOptions.RemoveEmptyEntries);
+                            if (splits.Count() == 2)
+                            {
+                                props.Add(Properties.Resources.StringLat, cdd.ToString(splits[0].Trim(), new CoordinateDDFormatter()));
+                                props.Add(Properties.Resources.StringLon, cdd.ToString("X" + splits[1].Trim(), new CoordinateDDFormatter()));
+                            }
+                            else
+                            {
+                                splits = output.Format.Split(new char[] { 'Y' }, StringSplitOptions.RemoveEmptyEntries);
+                                if (splits.Count() == 2)
+                                {
+                                    props.Add(Properties.Resources.StringLon, cdd.ToString(splits[0].Trim(), new CoordinateDDFormatter()));
+                                    props.Add(Properties.Resources.StringLat, cdd.ToString("Y" + splits[1].Trim(), new CoordinateDDFormatter()));
+                                }
+                                else
+                                {
+                                    props.Add(Properties.Resources.StringLat, cdd.Lat.ToString());
+                                    props.Add(Properties.Resources.StringLon, cdd.Lon.ToString());
+                                }
+                            }
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.DMS:
+                        CoordinateDMS cdms;
+                        if (coordinateGetter.CanGetDMS(output.SRFactoryCode, out coord) &&
+                            CoordinateDMS.TryParse(coord, out cdms))
+                        {
+                            output.OutputCoordinate = cdms.ToString(output.Format, new CoordinateDMSFormatter());
+                            var splits = output.Format.Split(new char[] { 'X' }, StringSplitOptions.RemoveEmptyEntries);
+                            if (splits.Count() == 2)
+                            {
+                                props.Add(Properties.Resources.StringLat, cdms.ToString(splits[0].Trim(), new CoordinateDMSFormatter()));
+                                props.Add(Properties.Resources.StringLon, cdms.ToString("X" + splits[1].Trim(), new CoordinateDMSFormatter()));
+                            }
+                            else
+                            {
+                                splits = output.Format.Split(new char[] { 'Y' }, StringSplitOptions.RemoveEmptyEntries);
+                                if (splits.Count() == 2)
+                                {
+                                    props.Add(Properties.Resources.StringLon, cdms.ToString(splits[0].Trim(), new CoordinateDMSFormatter()));
+                                    props.Add(Properties.Resources.StringLat, cdms.ToString("Y" + splits[1].Trim(), new CoordinateDMSFormatter()));
+                                }
+                                else
+                                {
+                                    props.Add(Properties.Resources.StringLat, cdms.ToString("A0°B0'C0.0\"N", new CoordinateDMSFormatter()));
+                                    props.Add(Properties.Resources.StringLon, cdms.ToString("X0°Y0'Z0.0\"E", new CoordinateDMSFormatter()));
+                                }
+                            }
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.DDM:
+                        CoordinateDDM ddm;
+                        if (coordinateGetter.CanGetDDM(output.SRFactoryCode, out coord) &&
+                            CoordinateDDM.TryParse(coord, out ddm))
+                        {
+                            output.OutputCoordinate = ddm.ToString(output.Format, new CoordinateDDMFormatter());
+                            var splits = output.Format.Split(new char[] { 'X' }, StringSplitOptions.RemoveEmptyEntries);
+                            if (splits.Count() == 2)
+                            {
+                                props.Add(Properties.Resources.StringLat, ddm.ToString(splits[0].Trim(), new CoordinateDDMFormatter()));
+                                props.Add(Properties.Resources.StringLon, ddm.ToString("X" + splits[1].Trim(), new CoordinateDDMFormatter()));
+                            }
+                            else
+                            {
+                                splits = output.Format.Split(new char[] { 'Y' }, StringSplitOptions.RemoveEmptyEntries);
+                                if (splits.Count() == 2)
+                                {
+                                    props.Add(Properties.Resources.StringLon, ddm.ToString(splits[0].Trim(), new CoordinateDDMFormatter()));
+                                    props.Add(Properties.Resources.StringLat, ddm.ToString("Y" + splits[1].Trim(), new CoordinateDDMFormatter()));
+                                }
+                                else
+                                {
+                                    props.Add(Properties.Resources.StringLat, ddm.ToString("A0°B0.0#####'N", new CoordinateDDMFormatter()));
+                                    props.Add(Properties.Resources.StringLon, ddm.ToString("X0°Y0.0#####'E", new CoordinateDDMFormatter()));
+                                }
+                            }
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.GARS:
+                        CoordinateGARS gars;
+                        if (coordinateGetter.CanGetGARS(output.SRFactoryCode, out coord) &&
+                            CoordinateGARS.TryParse(coord, out gars))
+                        {
+                            output.OutputCoordinate = gars.ToString(output.Format, new CoordinateGARSFormatter());
+                            props.Add(Properties.Resources.StringLon, gars.LonBand.ToString());
+                            props.Add(Properties.Resources.StringLat, gars.LatBand);
+                            props.Add(Properties.Resources.StringQuadrant, gars.Quadrant.ToString());
+                            props.Add(Properties.Resources.StringKey, gars.Key.ToString());
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.MGRS:
+                        CoordinateMGRS mgrs;
+                        if (coordinateGetter.CanGetMGRS(output.SRFactoryCode, out coord) &&
+                            CoordinateMGRS.TryParse(coord, out mgrs))
+                        {
+                            output.OutputCoordinate = mgrs.ToString(output.Format, new CoordinateMGRSFormatter());
+                            props.Add(Properties.Resources.StringGZD, mgrs.GZD);
+                            props.Add(Properties.Resources.StringGridSq, mgrs.GS);
+                            props.Add(Properties.Resources.StringEasting, mgrs.Easting.ToString("00000"));
+                            props.Add(Properties.Resources.StringNorthing, mgrs.Northing.ToString("00000"));
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.USNG:
+                        CoordinateUSNG usng;
+                        if (coordinateGetter.CanGetUSNG(output.SRFactoryCode, out coord) &&
+                            CoordinateUSNG.TryParse(coord, out usng))
+                        {
+                            output.OutputCoordinate = usng.ToString(output.Format, new CoordinateMGRSFormatter());
+                            props.Add(Properties.Resources.StringGZD, usng.GZD);
+                            props.Add(Properties.Resources.StringGridSq, usng.GS);
+                            props.Add(Properties.Resources.StringEasting, usng.Easting.ToString("00000"));
+                            props.Add(Properties.Resources.StringNorthing, usng.Northing.ToString("00000"));
+                            output.Props = props;
+                        }
+                        break;
+                    case CoordinateType.UTM:
+                        CoordinateUTM utm;
+                        if (coordinateGetter.CanGetUTM(output.SRFactoryCode, out coord) &&
+                            CoordinateUTM.TryParse(coord, out utm))
+                        {
+                            output.OutputCoordinate = utm.ToString(output.Format, new CoordinateUTMFormatter());
+                            props.Add(Properties.Resources.StringZone, utm.Zone.ToString() + utm.Hemi);
+                            props.Add(Properties.Resources.StringEasting, utm.Easting.ToString("000000"));
+                            props.Add(Properties.Resources.StringNorthing, utm.Northing.ToString("0000000"));
+                            output.Props = props;
+                        }
+                        break;
+                    default:
+                        break;
+                }
+            }
+        }
+
     }
 }
